scanlytic/requirements.txt
asgiref==3.8.1
azure-ai-documentintelligence==1.0.0
azure-core==1.32.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
idna==3.10
isodate==0.7.2
numpy==2.2.3
opencv-python==4.11.0.86
openpyxl==3.1.5
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==1.7.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virustotal-python==1.1.0
